requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
plotly>=5.24.0
openpyxl>=3.1.0
google-cloud-aiplatform>=1.114.0
python-dotenv>=1.0.0
